Made changes to Game Board System
Changes:

The value 2 is now used for a tetromino pieces that can't be moved, i.e. not the current piece and so can be moved.
This has made is a lot easier to distinguish previous tetromino pieces from the currently falling one, this has also allowed
for the second special case where a vertical I shaped tetromino drops down. 

To Do:
-Add functionality for moving pieces left and right as well as for rotation
-Clear lines and maybe a score system