Added use of jnativehook library
Java doesn't have a nice way to get keyboard input without having to use a library like Swing. Since this project uses the Console for the graphical output, Swing is unusable since it requires a component to be in focus from which input is captured. 
Jnativehook allows for obtaining/capturing the keyboard input that this project needs without having to use Swing and/or an ugly
component that serves no purpose but to capture input.
Java's Scanner requires the return key being used, Tetris blocks would have to be paused and a key entered every cycle, taking away the magic that makes Tetris.